+++ client/src/pages/financial-record/FaRecForm.tsx
@@ -1,32 +1,43 @@
 import styled from '@emotion/styled';
 import CommonStyles from '../../styles/CommonStyles';
+import axios from 'axios';
 import useInput from '../../hooks/useComponents';
 import PlusIcon from '../../../public/images/icon/plus.svg';
-import { FormEvent, useCallback, useState } from 'react';
-import axios from 'axios';
-import Cropper from 'react-easy-crop';
-import { Area, Point } from 'react-easy-crop/types';
-import { getCroppedImg } from '../../components/img-crop/canvasUtils';
-
-function readFile(file: File): Promise<string> {
-  return new Promise((resolve) => {
-    const reader = new FileReader();
-    reader.addEventListener(
-      'load',
-      () => {
-        if (typeof reader.result === 'string') {
-          resolve(reader.result);
-        }
-      },
-      false
-    );
-    reader.readAsDataURL(file);
-  });
-}
+import { FormEvent } from 'react';
+import { useImageCrop } from '../../hooks/useImgCrop';
+import { handleFileChange } from '../../components/img-crop/imgCropUtils';
+import ImgCropModal from '../../components/img-crop/ImgCropModal';
+import { useMutation } from '@tanstack/react-query';
+import { APIfinancialRecord } from '../../services/apiFinancial';
 
 export default function FaRecForm() {
   const [nameInput, faRecName] = useInput('text', '가계부 이름', 'faName');
   const [descInput, faRecDesc] = useInput('text', '가계부 설명', 'faDesc');
+  const {
+    imgSrc,
+    setImgSrc,
+    croppedImage,
+    setCroppedImage,
+    cropModal,
+    setCropModal,
+  } = useImageCrop();
+
+  const onFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
+    await handleFileChange({
+      e,
+      setCropModal,
+      setImgSrc,
+    });
+  };
+
+  const createRecordMutation = useMutation(APIfinancialRecord.createRecord, {
+    onSuccess: (data) => {
+      console.log('response-data', data);
+    },
+    onError: (error) => {
+      console.error(error);
+    },
+  });
 
   const handleSubmit = (e: FormEvent) => {
     e.preventDefault();
@@ -39,75 +50,34 @@ export default function FaRecForm() {
      */
     formData.append('userId', 'test');
 
-    axios
-      .post(
-        'https://zerohip-git-user-55-everland.vercel.app/api/financial-record/',
-        formData,
-        {
-          headers: {
-            'Content-Type': 'multipart/form-data',
-          },
-        }
-      )
-      .then((response) => {
-        console.log('response-data', response.data);
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
+    // axios
+    //   .post('/api/financial-record/', formData, {
+    //     headers: {
+    //       'Content-Type': 'multipart/form-data',
+    //     },
+    //   })
+    //   .then((response) => {
+    //     console.log('response-data', response.data);
+    //   })
+    //   .catch((error) => {
+    //     console.error(error);
+    //   });
 
-  // 크롭
-  const [imageSrc, setImageSrc] = useState<string | null>(null);
-  const [crop, setCrop] = useState<Point>({ x: 0, y: 0 });
-  const [zoom, setZoom] = useState(1);
-  const [croppedAreaPixels, setCroppedAreaPixels] = useState<Area | null>(null);
-  const [croppedImage, setCroppedImage] = useState<string | null>(null);
-  const [cropModal, setCropModal] = useState(false);
-
-  const onCropComplete = useCallback(
-    (croppedArea: Area, croppedAreaPixels: Area) => {
-      setCroppedAreaPixels(croppedAreaPixels);
-    },
-    []
-  );
-
-  const showCroppedImage = useCallback(async () => {
-    try {
-      if (imageSrc && croppedAreaPixels) {
-        const croppedImage = await getCroppedImg(imageSrc, croppedAreaPixels);
-        console.log(croppedImage); // Check the value
-        console.log(typeof croppedImage);
-        if (typeof croppedImage === 'string') {
-          setCroppedImage(croppedImage);
-        }
-        setCropModal(false);
-      }
-    } catch (e) {
-      console.error(e);
-    }
-  }, [imageSrc, croppedAreaPixels]);
-
-  const onFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files && e.target.files.length > 0) {
-      const file = e.target.files[0];
-      const imageDataUrl = await readFile(file);
-      setCropModal(true);
-      setImageSrc(imageDataUrl);
-    }
+    createRecordMutation.mutate(formData);
   };
 
   return (
     <S.Form onSubmit={handleSubmit}>
       <S.Container>
         <S.ImgBox>
-          <img
-            src={
-              croppedImage ||
-              'https://blog.kakaocdn.net/dn/bY6iW4/btrEJwN3Zrf/SqQZ605snSkSqP5U96S3AK/img.png'
-            }
-            alt=''
-          />
+          {croppedImage ? (
+            <img
+              src={croppedImage}
+              alt={faRecName ? `${faRecName} 프로필 사진` : '프로필 사진'}
+            />
+          ) : (
+            <img src='/images/icon/person.svg' alt='사람 아이콘' />
+          )}
 
           <S.FileInput
             type='file'
@@ -119,21 +89,15 @@ export default function FaRecForm() {
             <PlusIcon />
           </S.FileLabelBtn>
         </S.ImgBox>
-        {cropModal && imageSrc && (
-          <S.CropContainer>
-            <Cropper
-              image={imageSrc}
-              crop={crop}
-              zoom={zoom}
-              aspect={1 / 1}
-              onCropChange={setCrop}
-              onCropComplete={onCropComplete}
-              onZoomChange={setZoom}
-            />
-            <S.SubmitBtn type='button' onClick={showCroppedImage}>
-              이미지 선택 완료
-            </S.SubmitBtn>
-          </S.CropContainer>
+        {cropModal && (
+          <ImgCropModal
+            isOpen={cropModal}
+            setCropModal={setCropModal}
+            imgSrc={imgSrc}
+            aspect={1 / 1}
+            cropShape='round'
+            onCropComplete={(image) => setCroppedImage(image)}
+          />
         )}
         <S.InputFieldWrap>
           {nameInput}
@@ -160,11 +124,13 @@ const S = {
   `,
   Form: styled.form`
     width: 100%;
+    height: 100%;
     display: flex;
     align-items: center;
     justify-content: center;
   `,
   Container: styled.section`
+    position: relative;
     display: flex;
     flex-direction: column;
     align-items: center;
