+++ client/src/services/apiFinancial.ts
@@ -5,4 +5,12 @@ export const APIfinancialRecord = {
     const response = await axios.get('/data/recordList.json');
     return response.data;
   },
+  createRecord: async (formData) => {
+    const response = await axios.post('/api/financial-record/', formData, {
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      },
+    });
+    return response.data;
+  },
 };
